Airflow/requirements.txt
ultralytics
opencv-python-headless==4.9.0.80
numpy==1.26.4
pandas
pyzbar
pyzxing
matplotlib
psycopg2-binary
pyarrow